custom printf int
